requirements.txt
asgiref==3.7.2
Django==5.0
django-tinymce==3.6.1
Pillow==10.1.0
sqlparse==0.4.4
